% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Predict multiple new observations using a trained model}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{A trained model object (e.g., from randomForest) that supports \code{predict(..., type = "prob")}.}

\item{feature_df}{A data.frame of new observations. Must include columns: \code{gc_content}, \code{RNA_type}, \code{RNA_region}, \code{exon_length}, \code{distance_to_junction}, \code{evolutionary_conservation}, \code{DNA_5mer}.}

\item{positive_threshold}{Numeric. (default = 0.5) A threshold above which the predicted probability is labelled “Positive”; otherwise “Negative”.}
}
\value{
A data.frame with the same number of rows as \code{feature_df}, plus two new columns: \code{predicted_m6A_prob} (numeric) and \code{predicted_m6A_status} (character “Positive”/“Negative”).
}
\description{
This function uses a trained model object (\code{ml_fit}) to generate predictions on a new data frame (\code{feature_df}) of multiple observations. It enforces that certain required columns are present, recodes categorical variables, expands a 5‑mer DNA string into individual nucleotide‑position factors, and then uses \code{predict(..., type = "prob")} to compute the probability of the “Positive” class. It then assigns a status (“Positive” or “Negative”) based on the \code{positive_threshold}. The output is the original \code{feature_df} with two additional columns: \code{predicted_m6A_prob} and \code{predicted_m6A_status}.
}
\examples{
ml_model <- readRDS(system.file("extdata", "rf_fit.rds", package="m6APrediction"))
new_features <- read.csv(system.file("extdata", "m6A_input_example.csv",
                    package="m6APrediction"))
  result_df <- prediction_multiple(
    ml_fit = ml_model,
    feature_df = new_features,
    positive_threshold = 0.5
  )
  head(result_df)
}
